requirements.txt
discord-components
aiohttp
aiosignal
frozenlist
async-timeout
attrs
charset-normalizer
multidict
yarl
idna
discord.py==1.7.3
python-dotenv
openpyxl
cassiopeia
psycopg2-binary